Выделяет данные офферов и пунктов назначения в отдельные модели
