+++ src/mock/offers.js
@@ -1,4 +1,4 @@
-export const offers = [
+export const offersList = [
   {
     'type': 'taxi',
     'offers': [
